requirements.txt
Django==4.2.9
djangorestframework==3.14.0
PyMySQL==1.1.0
django-jazzmin==2.6.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.1